moved script out to sepearate file to prvent override in merge
